✨ split dynamic append patcher for loose and strict sandbox (#995)
* ✨ split dynamic append patcher for proxy and non-proxy sandbox

* ✨ change useLooseSandbox to sandbox.loose configuration

* ♿ improve warning message for non-proxy environment
