+++ examples/main/multiple.html
@@ -14,6 +14,10 @@
   </style>
 </head>
 <body>
+<button id="unmount">unmount</button>
+<button id="mount">mount</button>
+
+
 <div id="react15">react loading...</div>
 <div id="vue">vue loading...</div>
