+++ src/sandbox/index.ts
@@ -24,33 +24,26 @@ export { css } from './patchers';
  *
  * @param appName
  * @param elementGetter
- * @param singular
  * @param scopedCSS
+ * @param useLooseSandbox
  * @param excludeAssetFilter
  */
 export function createSandbox(
   appName: string,
   elementGetter: () => HTMLElement | ShadowRoot,
-  singular: boolean,
   scopedCSS: boolean,
+  useLooseSandbox?: boolean,
   excludeAssetFilter?: (url: string) => boolean,
 ) {
   let sandbox: SandBox;
   if (window.Proxy) {
-    sandbox = singular ? new LegacySandbox(appName) : new ProxySandbox(appName);
+    sandbox = useLooseSandbox ? new LegacySandbox(appName) : new ProxySandbox(appName);
   } else {
     sandbox = new SnapshotSandbox(appName);
   }
 
   // some side effect could be be invoked while bootstrapping, such as dynamic stylesheet injection with style-loader, especially during the development phase
-  const bootstrappingFreers = patchAtBootstrapping(
-    appName,
-    elementGetter,
-    sandbox,
-    singular,
-    scopedCSS,
-    excludeAssetFilter,
-  );
+  const bootstrappingFreers = patchAtBootstrapping(appName, elementGetter, sandbox, scopedCSS, excludeAssetFilter);
   // mounting freers are one-off and should be re-init at every mounting time
   let mountingFreers: Freer[] = [];
 
@@ -80,7 +73,7 @@ export function createSandbox(
 
       /* ------------------------------------------ 2. 开启全局变量补丁 ------------------------------------------*/
       // render 沙箱启动时开始劫持各类全局监听，尽量不要在应用初始化阶段有 事件监听/定时器 等副作用
-      mountingFreers = patchAtMounting(appName, elementGetter, sandbox, singular, scopedCSS, excludeAssetFilter);
+      mountingFreers = patchAtMounting(appName, elementGetter, sandbox, scopedCSS, excludeAssetFilter);
 
       /* ------------------------------------------ 3. 重置一些初始化时的副作用 ------------------------------------------*/
       // 存在 rebuilder 则表明有些副作用需要重建
