+++ src/sandbox/patchers/dynamicAppend/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @author Kuitos
+ * @since 2020-10-13
+ */
+
+export { patchLooseSandbox } from './forLooseSandbox';
+export { patchStrictSandbox } from './forStrictSandbox';
