+++ src/sandbox/patchers/index.ts
@@ -4,18 +4,16 @@
  */
 
 import { Freer, SandBox, SandBoxType } from '../../interfaces';
-import patchDynamicAppend from './dynamicAppend';
+import * as css from './css';
+import { patchLooseSandbox, patchStrictSandbox } from './dynamicAppend';
 import patchHistoryListener from './historyListener';
 import patchInterval from './interval';
 import patchWindowListener from './windowListener';
-
-import * as css from './css';
 
 export function patchAtMounting(
   appName: string,
   elementGetter: () => HTMLElement | ShadowRoot,
   sandbox: SandBox,
-  singular: boolean,
   scopedCSS: boolean,
   excludeAssetFilter?: Function,
 ): Freer[] {
@@ -23,13 +21,21 @@ export function patchAtMounting(
     () => patchInterval(sandbox.proxy),
     () => patchWindowListener(sandbox.proxy),
     () => patchHistoryListener(),
-    () => patchDynamicAppend(appName, elementGetter, sandbox.proxy, true, singular, scopedCSS, excludeAssetFilter),
   ];
 
   const patchersInSandbox = {
-    [SandBoxType.LegacyProxy]: [...basePatchers],
-    [SandBoxType.Proxy]: [...basePatchers],
-    [SandBoxType.Snapshot]: basePatchers,
+    [SandBoxType.LegacyProxy]: [
+      ...basePatchers,
+      () => patchLooseSandbox(appName, elementGetter, sandbox.proxy, true, scopedCSS, excludeAssetFilter),
+    ],
+    [SandBoxType.Proxy]: [
+      ...basePatchers,
+      () => patchStrictSandbox(appName, elementGetter, sandbox.proxy, true, scopedCSS, excludeAssetFilter),
+    ],
+    [SandBoxType.Snapshot]: [
+      ...basePatchers,
+      () => patchLooseSandbox(appName, elementGetter, sandbox.proxy, true, scopedCSS, excludeAssetFilter),
+    ],
   };
 
   return patchersInSandbox[sandbox.type]?.map(patch => patch());
@@ -39,18 +45,19 @@ export function patchAtBootstrapping(
   appName: string,
   elementGetter: () => HTMLElement | ShadowRoot,
   sandbox: SandBox,
-  singular: boolean,
   scopedCSS: boolean,
   excludeAssetFilter?: Function,
 ): Freer[] {
-  const basePatchers = [
-    () => patchDynamicAppend(appName, elementGetter, sandbox.proxy, false, singular, scopedCSS, excludeAssetFilter),
-  ];
-
   const patchersInSandbox = {
-    [SandBoxType.LegacyProxy]: basePatchers,
-    [SandBoxType.Proxy]: basePatchers,
-    [SandBoxType.Snapshot]: basePatchers,
+    [SandBoxType.LegacyProxy]: [
+      () => patchLooseSandbox(appName, elementGetter, sandbox.proxy, false, scopedCSS, excludeAssetFilter),
+    ],
+    [SandBoxType.Proxy]: [
+      () => patchStrictSandbox(appName, elementGetter, sandbox.proxy, false, scopedCSS, excludeAssetFilter),
+    ],
+    [SandBoxType.Snapshot]: [
+      () => patchLooseSandbox(appName, elementGetter, sandbox.proxy, false, scopedCSS, excludeAssetFilter),
+    ],
   };
 
   return patchersInSandbox[sandbox.type]?.map(patch => patch());
